update plist_spec.js, test not work yet, ubicall/node-red-contrib-ivr#28
